pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-mesh"
version = "0.1.0"
description = "Lightweight multi-agent coordination runtime using the A2A protocol"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Sushaan Kandukoori", email = "sushaan@users.noreply.github.com" },
]
keywords = ["a2a", "agents", "multi-agent", "coordination", "runtime", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "httpx>=0.25.0",
    "click>=8.1",
    "pyjwt>=2.8",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "structlog>=23.1",
    "uvicorn>=0.24",
    "starlette>=0.27",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1.0",
    "mypy>=1.7",
    "respx>=0.20",
    "httpx",
]

[project.scripts]
a2a-mesh = "a2a_mesh.cli:cli"

[project.urls]
Homepage = "https://github.com/sushaan-k/a2a-mesh"
Repository = "https://github.com/sushaan-k/a2a-mesh"
Issues = "https://github.com/sushaan-k/a2a-mesh/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_mesh"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["a2a_mesh"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
